+++ doc/history.md
@@ -3,7 +3,10 @@
  + Fix nested emphasis
  + Add some basic edit operations to text input widget
  + Add inline links [<dest>]:#<name>
-
 ### 1/22/2017 ###
  + Add more toolbar functions
  + Improve syntax highlighting
+### 12/16/2017 ###
+ + Make reference markdown widgets optional
+ + Add widget object names
+ + Support fixed font for editor